refactor: 리포지토리 추상화 완료
