+++ src/modules/base/auth/core/auth.service.ts
@@ -2,9 +2,9 @@ import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
 import Redis from 'ioredis';
 import * as bcrypt from 'bcrypt';
 import { InjectRedis } from '@nestjs-modules/ioredis';
-import { ToolService } from '../../../utils/tool/tool.service';
-import { MailService } from '../../../utils/mail/mail.service';
-import { NicknameService } from '../../../utils/nickname/nickname.service';
+import { ToolService } from '../../../utils/tool/core/tool.service';
+import { MailService } from '../../../utils/mail/core/mail.service';
+import { NicknameService } from '../../../utils/nickname/core/nickname.service';
 import { AuthRepository } from '../infrastructure/auth.repository';
 import { CreateUserReqDto } from '../dto/request/createUserReq.dto';
 import { OauthDto } from '../dto/oauth.dto';
