+++ src/modules/extensions/essay/badge/core/badge.service.ts
@@ -1,6 +1,6 @@
 import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
 import { Transactional } from 'typeorm-transactional';
-import { ToolService } from '../../../../utils/tool/tool.service';
+import { ToolService } from '../../../../utils/tool/core/tool.service';
 import { BadgeRepository } from '../infrastructure/badge.repository';
 import { Tag } from '../../../../../entities/tag.entity';
 import { User } from '../../../../../entities/user.entity';
