+++ src/modules/extensions/essay/story/core/story.service.ts
@@ -9,7 +9,7 @@ import {
 import { User } from '../../../../../entities/user.entity';
 import { Story } from '../../../../../entities/story.entity';
 import { StoryDto } from '../dto/story.dto';
-import { ToolService } from '../../../../utils/tool/tool.service';
+import { ToolService } from '../../../../utils/tool/core/tool.service';
 import { UserService } from '../../../../base/user/core/user.service';
 import { CreateStoryReqDto } from '../dto/repuest/createStoryReq.dto';
 import { EssayService } from '../../../../base/essay/core/essay.service';
