+++ src/modules/utils/cron/cron.module.ts
@@ -3,12 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../../../entities/user.entity';
 import { Essay } from '../../../entities/essay.entity';
 import { CronLog } from '../../../entities/cronLog.entity';
-import { CronService } from './cron.service';
+import { CronService } from './core/cron.service';
 import { BullModule } from '@nestjs/bull';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { DeactivationReason } from '../../../entities/deactivationReason.entity';
 import { EssayModule } from '../../base/essay/essay.module';
-import { CronProcessor } from './cron.processor';
+import { CronProcessor } from './core/cron.processor';
 import { Geulroquis } from '../../../entities/geulroguis.entity';
 import { ScheduleModule } from '@nestjs/schedule';
 import { ToolModule } from '../tool/tool.module';
