+++ src/modules/utils/mail/mail.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MailService } from './mail.service';
+import { MailService } from './core/mail.service';
 import { ConfigModule } from '@nestjs/config';
 
 @Module({
