+++ src/modules/utils/nickname/nickname.module.ts
@@ -1,12 +1,12 @@
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module } from '@nestjs/common';
-import { NicknameService } from './nickname.service';
-import { NicknameRepository } from './nickname.repository';
+import { NicknameService } from './core/nickname.service';
+import { NicknameRepository } from './infrastructure/nickname.repository';
 import { BasicNickname } from '../../../entities/basicNickname.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([BasicNickname])],
-  providers: [NicknameService, NicknameRepository],
-  exports: [NicknameService, NicknameRepository],
+  providers: [NicknameService, { provide: 'INicknameRepository', useClass: NicknameRepository }],
+  exports: [NicknameService, { provide: 'INicknameRepository', useClass: NicknameRepository }],
 })
 export class NicknameModule {}
